requirements.txt
Django==3.1.6
arrow
djangorestframework>=3.11
django-cors-headers
six
waitress
psycopg2
djangorestframework-simplejwt